refactoring WS ref processor to separate package
